Moved transfer with depth map algo implementation to core arrow
