+++ arrows/core/CMakeLists.txt
@@ -44,6 +44,7 @@ set( plugin_core_headers
   render_mesh_depth_map.h
   track_features_augment_keyframes.h
   track_features_core.h
+  transfer_with_depth_map.h
   transform.h
   transform_detected_object_set.h
   triangle_scan_iterator.h
@@ -126,6 +127,7 @@ set( plugin_core_sources
   render_mesh_depth_map.cxx
   track_features_augment_keyframes.cxx
   track_features_core.cxx
+  transfer_with_depth_map.cxx
   transform.cxx
   transform_detected_object_set.cxx
   triangle_scan_iterator.cxx
