+++ arrows/core/tests/data/dest_camera.krtd
