+++ arrows/core/tests/data/src_camera.krtd
