+++ doc/release-notes/master.txt
@@ -96,6 +96,11 @@ Arrows: Core
    transforming detected object sets from one camera perspective to another.
    Implements vital::algo::detected_object_filter.
 
+ * Added an algorithm implementation (transfer_with_depth_map) for
+   transforming detected object sets from one camera perspective to
+   another with respect to a depth map.  Implements
+   vital::algo::detected_object_filter.
+
  * Added a detect_features adaptor that applies a filter_features
    algorithm to filter features returned by the detector.  This allows
    a filter to be injected anywhere a detector is normally used.
@@ -197,11 +202,6 @@ Arrows: GDAL
 Arrows: KPF
 
 Arrows: OpenCV
-
- * Added an algorithm implementation (transfer_with_depth_map) for
-   transforming detected object sets from one camera perspective to
-   another with respect to a depth map.  Implements
-   vital::algo::detected_object_filter.
 
 Arrows: Serialization
